Change to dollar icon
