--- src/components/icons/Rocket.tsx
@@ -1,40 +1,0 @@
-export default function () {
-  return (
-    <svg viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <g clip-path="url(#clip0_257_2133)">
-        <mask
-          id="mask0_257_2133"
-          style="mask-type:luminance"
-          maskUnits="userSpaceOnUse"
-          x="0"
-          y="0"
-          width="16"
-          height="16"
-        >
-          <path d="M16 0H0V16H16V0Z" fill="white" />
-        </mask>
-        <g mask="url(#mask0_257_2133)">
-          <path
-            fill-rule="evenodd"
-            clip-rule="evenodd"
-            d="M11.1006 1.22969C9.96346 1.29611 8.88644 1.77798 8.08099 2.58344L3.71686 6.94757C2.58062 8.08382 2.57256 9.92353 3.69883 11.0697C3.84136 11.2147 4.07482 11.2157 4.21862 11.0719L5.95754 9.33304C6.1528 9.13777 6.46938 9.13777 6.66464 9.33304C6.85991 9.52829 6.85991 9.84488 6.66464 10.0401L4.92574 11.7791C4.78194 11.9229 4.78295 12.1563 4.92802 12.2989C6.07419 13.4252 7.91399 13.4172 9.05027 12.2808L13.4323 7.89876C14.2267 7.10433 14.7055 6.04832 14.7795 4.92725C14.917 2.84165 13.1872 1.10782 11.1006 1.22969ZM11.333 5.99937C11.333 6.73576 10.7361 7.3327 9.99968 7.3327C9.26331 7.3327 8.66635 6.73576 8.66635 5.99937C8.66635 5.263 9.26331 4.66604 9.99968 4.66604C10.7361 4.66604 11.333 5.263 11.333 5.99937Z"
-            fill="currentColor"
-          />
-          <path
-            d="M1.40458 6.17039C0.793429 5.76143 0.708488 4.89589 1.22848 4.3759L1.47826 4.12614C2.52586 3.07857 4.06013 2.68334 5.48346 3.09438C5.85535 3.20178 5.97566 3.66878 5.70195 3.9425L3.16373 6.48071C3.03854 6.6059 2.97595 6.66849 2.90895 6.70963C2.73446 6.81677 2.52042 6.83777 2.32842 6.76661C2.25471 6.73927 2.18114 6.69005 2.03401 6.59159L1.40458 6.17039Z"
-            fill="currentColor"
-          />
-          <path
-            d="M9.74016 13.6103C9.64158 13.4631 9.59231 13.3894 9.56496 13.3156C9.49387 13.1237 9.51485 12.9098 9.62189 12.7354C9.66305 12.6683 9.72571 12.6056 9.85103 12.4803L12.3891 9.9422C12.6629 9.66849 13.1298 9.78875 13.2373 10.1606C13.6486 11.5842 13.2534 13.1189 12.2056 14.1667L11.956 14.4163C11.4361 14.9362 10.5707 14.8513 10.1617 14.2402L9.74016 13.6103Z"
-            fill="currentColor"
-          />
-        </g>
-      </g>
-      <defs>
-        <clipPath id="clip0_257_2133">
-          <rect width="16" height="16" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
-  )
-}
